fix: ssl:make timing bug (#180)
createCA() is async and thus doesn't finish till *after* it returns.
createCert() *needs* the CA file to exist before it runs.

If both files are missing, this command died every time because
`createCert()` would be called before the CA file was created.

It would then kill the node process before *either* file was created.

Now we use async/await to order them appropriately and we add a test to
ensure the files get created.